refactor(SB-1111) - move npm commands to utility
