+++ package.json
@@ -28,6 +28,7 @@
     "chalk": "^2.1.0",
     "commander": "^2.11.0",
     "config": "^1.26.2",
+    "execa": "^0.10.0",
     "fs-extra": "^5.0.0",
     "gunzip-maybe": "^1.4.1",
     "hostile": "^1.3.0",
